+++ CHANGES.rst
@@ -5,7 +5,21 @@ Changelog of lizard-auth-server
 1.8 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- A user arriving at the SSO server after being redirected there can
+  now use a "return_unauthenticated" URL attribute. If the user is
+  already logged in on the SSO server, redirects are set up so that he
+  will be logged in on the site he was redirected from.
+
+  If he is not, then if return_unauthenticated is False (the default,
+  and the old behaviour), then he will be forced to log in before
+  being redirected back.
+
+  If return_unauthenticated, redirect the user back without logging in
+  (to lizard-auth-client's /sso/local_not_logged_in/ URL).
+
+  This enables a "attempt to auto-login if possible, but don't require it"
+  workflow that is sometimes helpful.
+
 
 
 1.7 (2016-06-14)