+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog of lizard-auth-server
 1.8 (unreleased)
 ----------------
 
+- Renamed 'return_unauthenticated' to 'attempt_login_if_authenticated'.
+
 - A user arriving at the SSO server after being redirected there can
   now use a "return_unauthenticated" URL attribute. If the user is
   already logged in on the SSO server, redirects are set up so that he